Flag new go1.8 features
It appears being able to return a redirect response itself uses new
functionality in go 1.8. So separate that code out and add a build flag.
Only affects tests.
