+++ repo_test.go
@@ -11,17 +11,10 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-var (
-	datasource *purldb
-)
-
 func TestAllPurls(t *testing.T) {
 	assert := assert.New(t)
 
-	var purltestdb *purldb
-	purltestdb = datasource
-
-	result := purltestdb.AllPurls()
+	result := datasource.AllPurls()
 
 	for _, res := range result {
 		assert.NotEqual(res.Date_created, time.Time{}, "Time incorrectly set on repo")
@@ -32,9 +25,7 @@ func TestAllPurls(t *testing.T) {
 func TestFindPurl(t *testing.T) {
 	assert := assert.New(t)
 
-	var purltestdb *purldb
-	purltestdb = datasource
-	result := purltestdb.FindPurl(5)
+	result := datasource.FindPurl(5)
 
 	assert.NotEqual(result.Date_created, time.Time{}, "Time incorrectly set on repo")
 	assert.NotEqual(result.Id, nil, "Id nil")
@@ -42,57 +33,40 @@ func TestFindPurl(t *testing.T) {
 	assert.Equal(result.Id, 5, "Id not correct")
 	assert.Equal(result.Repo_obj_id, "5", "Repo Id not correct")
 	assert.Equal(result.Access_count, 625, "Id not correct")
-	time_val, err := time.Parse(time.RFC3339, "2016-11-15T14:16:14Z")
-	if err != nil {
-		panic(err)
-	}
+	time_val, _ := time.Parse(time.RFC3339, "2016-11-15T14:16:14Z")
 	assert.Equal(result.Last_accessed, time_val, "Last_accesed not correct")
-	time_val, err = time.Parse(time.RFC3339, "2011-09-14T14:40:11Z")
-	if err != nil {
-		panic(err)
-	}
+	time_val, _ = time.Parse(time.RFC3339, "2011-09-14T14:40:11Z")
 	assert.Equal(result.Date_created, time_val, "Date_created not correct")
 }
 
 func TestCreatePurl(t *testing.T) {
 	assert := assert.New(t)
 
-	var newpurl Purl
-	newpurl.Id = 11
-	newpurl.Repo_obj_id = "110"
-	var err error
-	newpurl.Last_accessed, err = time.Parse(time.RFC3339, "2016-11-16T03:33:33Z")
-	if err != nil {
-		panic(err)
+	var newpurl = Purl{
+		Id:          11,
+		Repo_obj_id: "110",
 	}
-	newpurl.Date_created, err = time.Parse(time.RFC3339, "2011-09-14T13:55:55Z")
-	if err != nil {
-		panic(err)
-	}
+	newpurl.Last_accessed, _ = time.Parse(time.RFC3339, "2016-11-16T03:33:33Z")
+	newpurl.Date_created, _ = time.Parse(time.RFC3339, "2011-09-14T13:55:55Z")
 
-	var purltestdb *purldb
-	purltestdb = datasource
-	_ = purltestdb.createPurlDB(newpurl)
+	datasource.CreatePurl(newpurl)
 
-	result := purltestdb.FindPurl(11)
+	result := datasource.FindPurl(11)
 
 	assert.Equal(result.Id, newpurl.Id, "Id not correct")
 	assert.Equal(result.Repo_obj_id, newpurl.Repo_obj_id, "Repo Id not correct")
 	assert.Equal(result.Last_accessed, newpurl.Last_accessed, "Last_accesed not correct")
 	assert.Equal(result.Date_created, newpurl.Date_created, "Date_created not correct")
 
-	_ = purltestdb.destroyPurlDB(11)
+	// _ = datasource.destroyPurlDB(11)
 }
 
 func init() {
+	connection := os.Getenv("MYSQL_CONNECTION")
+	if connection == "" {
+		connection = "/test"
+		fmt.Println("MYSQL_CONNECTION not set. Using default:", connection)
+	}
 
-	mysqlLocation := fmt.Sprintf("%s:%s@tcp(%s:%s)/%s",
-		os.Getenv("MYSQL_USER"),
-		os.Getenv("MYSQL_PASSWORD"),
-		os.Getenv("MYSQL_HOST"),
-		os.Getenv("MYSQL_PORT"),
-		os.Getenv("MYSQL_DB"),
-	)
-
-	datasource = NewDBSource(mysqlLocation)
+	datasource = NewDBSource(connection)
 }
